refactor(weave): extract monitor query validation to its own module
Addresses review feedback: moves the monitor-query validation, error-message
construction, and constants out of the 3.6k-line calls_query_builder.py into a
sibling monitor_query_validation.py (calls_query_builder is now untouched).
Names the two guards (_is_monitor_object, _monitor_query) and renames the
field check to validate_calls_query with a docstring that matches behavior:
it validates by compiling, surfacing both bad field refs and malformed structure.
