+++ weave/trace_server/clickhouse_trace_server_batched.py
@@ -103,6 +103,8 @@ from weave.trace_server.calls_query_builder.calls_query_builder import (
     build_calls_complete_update_query,
     build_calls_stats_query,
     combine_conditions,
+)
+from weave.trace_server.calls_query_builder.monitor_query_validation import (
     validate_monitor_query_fields,
 )
 from weave.trace_server.calls_query_builder.usage_query_builder import (
